specify exceptions, move custom exception, use list comp for None model_ids
