+++ src/triage/component/catwalk/exceptions.py
