+++ src/triage/experiments/multicore.py
@@ -66,7 +66,7 @@ class MultiCoreExperiment(ExperimentBase):
             ):
                 model_ids += batch_model_ids
             # remove Nones from model_ids so we don't try to predict/evaluate on skipped baselines
-            model_ids = list(filter(None.__ne__, model_ids))
+            model_ids = [model_id for model_id in model_ids if model_id is not None]
             logging.info('Done training models')
 
             for split_def, test_uuid in zip(
